+++ src/components/BulkEditModal/index.ts
@@ -1,0 +1,1 @@
+export { default as BulkEditModal } from './BulkEditModal';