+++ src/components/CheckList/CheckList.tsx
@@ -27,11 +27,12 @@ import {
   RadioButtonGroup,
   InlineSwitch,
   AsyncMultiSelect,
+  ButtonCascader,
 } from '@grafana/ui';
 import { unEscapeStringFromRegex, escapeStringForRegex, GrafanaTheme, AppEvents, SelectableValue } from '@grafana/data';
 import { matchesAllFilters } from './checkFilters';
 import {
-  fetchProbes,
+  fetchProbeOptions,
   deleteSelectedChecks,
   deleteSingleCheck,
   getIconOverlayToggleFromLS,
@@ -57,6 +58,7 @@ import { SuccessRateContext, SuccessRateTypes } from 'contexts/SuccessRateContex
 import { ChecksVisualization } from '../ChecksVisualization';
 import ThresholdGlobalSettings from '../Thresholds/ThresholdGlobalSettings';
 import { useNavigation } from 'hooks/useNavigation';
+import { BulkEditModal } from 'components/BulkEditModal';
 
 const getStyles = (theme: GrafanaTheme) => ({
   headerContainer: css`
@@ -144,6 +146,7 @@ export const CheckList = ({ instance, checks, onCheckUpdate }: Props) => {
   const [bulkActionInProgress, setBulkActionInProgress] = useState(false);
 
   const [showThresholdModal, setShowThresholdModal] = useState(false);
+  const [bulkEditAction, setBulkEditAction] = useState<'add' | 'remove' | null>(null);
   const navigate = useNavigation();
 
   const styles = useStyles(getStyles);
@@ -418,7 +421,7 @@ export const CheckList = ({ instance, checks, onCheckUpdate }: Props) => {
             });
           }}
           defaultOptions
-          loadOptions={() => fetchProbes(instance)}
+          loadOptions={() => fetchProbeOptions(instance)}
           value={checkFilters.probes}
           placeholder="All probes"
           allowCustomValue={false}
@@ -447,18 +450,29 @@ export const CheckList = ({ instance, checks, onCheckUpdate }: Props) => {
                   Select all {filteredChecks.length} checks
                 </Button>
               )}
-              <Button
-                type="button"
-                variant="destructive"
-                className={styles.marginRightSmall}
-                onClick={handleDeleteSelectedChecks}
-                disabled={!hasRole(OrgRole.EDITOR) || bulkActionInProgress}
-              >
-                Delete
-              </Button>
+              {selectedChecks.size > 1 && (
+                <ButtonCascader
+                  options={[
+                    {
+                      label: 'Add probes',
+                      value: 'add',
+                    },
+                    {
+                      label: 'Remove probes',
+                      value: 'remove',
+                    },
+                  ]}
+                  className={styles.marginRightSmall}
+                  disabled={!hasRole(OrgRole.EDITOR) || bulkActionInProgress}
+                  onChange={(value) => setBulkEditAction(value[0] as any)}
+                >
+                  Bulk Edit Probes
+                </ButtonCascader>
+              )}
               <Button
                 type="button"
                 variant="primary"
+                fill="text"
                 onClick={handleEnableSelectedChecks}
                 className={styles.marginRightSmall}
                 disabled={!hasRole(OrgRole.EDITOR) || bulkActionInProgress}
@@ -468,10 +482,23 @@ export const CheckList = ({ instance, checks, onCheckUpdate }: Props) => {
               <Button
                 type="button"
                 variant="secondary"
+                fill="text"
                 onClick={handleDisableSelectedChecks}
+                className={styles.marginRightSmall}
                 disabled={!hasRole(OrgRole.EDITOR) || bulkActionInProgress}
               >
                 Disable
+              </Button>
+
+              <Button
+                type="button"
+                variant="destructive"
+                fill="text"
+                className={styles.marginRightSmall}
+                onClick={handleDeleteSelectedChecks}
+                disabled={!hasRole(OrgRole.EDITOR) || bulkActionInProgress}
+              >
+                Delete
               </Button>
             </div>
           </>
@@ -553,6 +580,20 @@ export const CheckList = ({ instance, checks, onCheckUpdate }: Props) => {
         onError={() =>
           appEvents.emit(AppEvents.alertError, [`Error updating thresholds. make sure your values don't overlap`])
         }
+      />
+      <BulkEditModal
+        instance={instance}
+        selectedChecks={getChecksFromSelected}
+        onDismiss={() => setBulkEditAction(null)}
+        action={bulkEditAction}
+        isOpen={bulkEditAction !== null}
+        onSuccess={() => {
+          onCheckUpdate();
+          appEvents.emit(AppEvents.alertSuccess, ['All selected checks successfully updated']);
+        }}
+        onError={(err) => {
+          appEvents.emit(AppEvents.alertError, [`There was an error updating checks: ${err}`]);
+        }}
       />
     </div>
   );