+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+# [1.9.3](https://github.com/grafana/synthetic-monitoring-app/compare/v1.9.2...v1.9.3) (2022-1-10)
+
+### Bug fixes
+
+- Gracefully handle inability to connect to API
+
+# [1.9.2](https://github.com/grafana/synthetic-monitoring-app/compare/v1.8.1...v1.9.2) (2022-1-06)
+
+### Features
+
+- Added the ability to bulk edit probes for checks
+
 # [1.8.1](https://github.com/grafana/synthetic-monitoring-app/compare/v1.8.0...v1.8.1) (2022-1-04)
 
 ### Bug Fixes